reorganize routing for main component
