+++ client/src/App.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Nav from "./components/Nav";
+import Main from "./components/Main";
+import { Route, Link } from "react-router-dom";
+
+const App = () =>
+  <div>
+    <Nav />
+    <Main />
+  </div>;
+
+export default App;
