+++ client/src/index.js
@@ -1,5 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Main from "./Main";
+import Main from "./components/Main";
+import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(<Main />, document.getElementById("root"));
+ReactDOM.render(
+  <BrowserRouter>
+    <Main />
+  </BrowserRouter>, document.getElementById("root")
+);
